Some reworks on script processing.
Action "parse" added.
Redesigned "list", "copy" and "unzip" actions.
Changed template patern of lists to {#...{@...@}...#}.
Changed templates' default encoding to ANSI.

+++ SourceGen/ActionsMethods.cs
@@ -1,4 +1,11 @@
-﻿using System;
+﻿/*
+ * Copyright (C) 2018 Maxim Yudin <i@hal.su>. All rights reserved.
+ * 
+ * This file is a part of the closed source section of LOCM3Gen project.
+ * You may NOT use, distribute, copy or modify this file without special author's permission.
+ */
+
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.IO.Compression;
@@ -54,24 +61,16 @@ namespace SourceGen
             var value = ParseString(nestedElement.Attribute("value")?.Value?.Trim() ?? "");
             list.Remove(value);
           }
-          else if (childActionName == "add-files")
+          else if (childActionName == "add-paths")
           {
-            var sourcePattern = ParseString(nestedElement.Attribute("source")?.Value?.Trim() ?? "");
-            if (sourcePattern != "")
+            var sourcePath = ParseString(nestedElement.Attribute("source-dir")?.Value?.Trim() ?? "");
+            if (sourcePath != "" && Directory.Exists(sourcePath))
             {
-              var sourcePath = Path.GetFullPath(Path.GetDirectoryName(sourcePattern));
-              if (Directory.Exists(sourcePath))
-              {
-                var fileNamePattern = Path.GetFileName(sourcePattern);
-                var recursive = (nestedElement.Attribute("recursive")?.Value?.Trim()?.ToLower() ?? "") == "true";
-                var relative = (nestedElement.Attribute("relative")?.Value?.Trim()?.ToLower() ?? "") == "true";
-                foreach (var fileName in Directory.EnumerateFiles(sourcePath, fileNamePattern, recursive ? SearchOption.AllDirectories : SearchOption.TopDirectoryOnly))
-                  list.Add(relative ? fileName.Replace(sourcePath, ".") : fileName);
-              }
-              else
-              {
-                // TODO: Wrong source path.
-              }
+              var fileNamePattern = nestedElement.Attribute("file-pattern")?.Value?.Trim() ?? "";
+              var recursive = (nestedElement.Attribute("recursive")?.Value?.Trim()?.ToLower() ?? "") == "true";
+              var relative = (nestedElement.Attribute("relative")?.Value?.Trim()?.ToLower() ?? "") == "true";
+              foreach (var fileName in Directory.EnumerateFiles(sourcePath, fileNamePattern, recursive ? SearchOption.AllDirectories : SearchOption.TopDirectoryOnly))
+                list.Add(relative ? fileName.Replace(sourcePath, ".") : fileName);
             }
             else
             {
@@ -98,34 +97,26 @@ namespace SourceGen
     [ActionNames("copy")]
     public void CopyAction(XElement actionElement)
     {
-      var sourcePattern = ParseString(actionElement.Attribute("source")?.Value?.Trim() ?? "");
-      if (sourcePattern != "")
+      var sourcePath = ParseString(actionElement.Attribute("source-dir")?.Value?.Trim() ?? "");
+      if (sourcePath != "" && Directory.Exists(sourcePath))
       {
-        var sourcePath = Path.GetFullPath(Path.GetDirectoryName(sourcePattern));
-        if (Directory.Exists(sourcePath))
+        var fileNamePattern = actionElement.Attribute("file-pattern")?.Value?.Trim() ?? "";
+        var destinationPath = Path.GetFullPath(ParseString(actionElement.Attribute("target-dir")?.Value?.Trim() ?? ""));
+        var recursive = (actionElement.Attribute("recursive")?.Value?.Trim()?.ToLower() ?? "") == "true";
+        var keepExistingFiles = (actionElement.Attribute("keep-existing")?.Value?.Trim()?.ToLower() ?? "") == "true";
+        var parseFiles = (actionElement.Attribute("parse")?.Value?.Trim()?.ToLower() ?? "") == "true";
+        foreach (var fileName in Directory.EnumerateFiles(sourcePath, fileNamePattern, recursive ? SearchOption.AllDirectories : SearchOption.TopDirectoryOnly))
         {
-          var fileNamePattern = Path.GetFileName(sourcePattern);
-          var destinationPath = Path.GetFullPath(ParseString(actionElement.Attribute("destination")?.Value?.Trim() ?? ""));
-          var recursive = (actionElement.Attribute("recursive")?.Value?.Trim()?.ToLower() ?? "") == "true";
-          var keepExistingFiles = (actionElement.Attribute("keep-existing")?.Value?.Trim()?.ToLower() ?? "") == "true";
-          var parseFiles = (actionElement.Attribute("parse")?.Value?.Trim()?.ToLower() ?? "") == "true";
-          foreach (var fileName in Directory.EnumerateFiles(sourcePath, fileNamePattern, recursive ? SearchOption.AllDirectories : SearchOption.TopDirectoryOnly))
+          var destinationFileName = fileName.Replace(sourcePath, destinationPath);
+          var destinationDirectory = Path.GetDirectoryName(destinationFileName);
+          if (!Directory.Exists(destinationDirectory))
+            Directory.CreateDirectory(destinationDirectory);
+          if (!File.Exists(destinationFileName) || !keepExistingFiles)
           {
-            var destinationFileName = fileName.Replace(sourcePath, destinationPath);
-            var destinationDirectory = Path.GetDirectoryName(destinationFileName);
-            if (!Directory.Exists(destinationDirectory))
-              Directory.CreateDirectory(destinationDirectory);
-            if (!File.Exists(destinationFileName) || !keepExistingFiles)
-            {
-              File.Copy(fileName, destinationFileName, true);
-              if (parseFiles)
-                ParseFile(destinationFileName);
-            }
+            File.Copy(fileName, destinationFileName, true);
+            if (parseFiles)
+              ParseFile(destinationFileName);
           }
-        }
-        else
-        {
-          // TODO: Wrong source path.
         }
       }
       else
@@ -150,7 +141,7 @@ namespace SourceGen
           var svdStream = svdArchive.GetEntry(entryPath);
           if (svdStream != null)
           {
-            var destinationPath = Path.GetFullPath(ParseString(actionElement.Attribute("destination")?.Value?.Trim() ?? ""));
+            var destinationPath = Path.GetFullPath(ParseString(actionElement.Attribute("target-dir")?.Value?.Trim() ?? ""));
             var destinationFileName = Path.Combine(destinationPath, Path.GetFileName(entryPath));
             var keepExistingFiles = (actionElement.Attribute("keep-existing")?.Value?.Trim()?.ToLower() ?? "") == "true";
             if (!Directory.Exists(destinationPath))
@@ -173,6 +164,27 @@ namespace SourceGen
       else
       {
         // TODO: Wrong archive path.
+      }
+    }
+
+    /// <summary>
+    /// Action method for extracting a file from a zip archive.
+    /// </summary>
+    /// <param name="data">XML element containing action data.</param>
+    [ActionNames("parse")]
+    public void ParseAction(XElement actionElement)
+    {
+      var sourcePath = ParseString(actionElement.Attribute("source-dir")?.Value?.Trim() ?? "");
+      if (sourcePath != "" && Directory.Exists(sourcePath))
+      {
+        var fileNamePattern = actionElement.Attribute("file-pattern")?.Value?.Trim() ?? "";
+        var recursive = (actionElement.Attribute("recursive")?.Value?.Trim()?.ToLower() ?? "") == "true";
+        foreach (var fileName in Directory.EnumerateFiles(sourcePath, fileNamePattern, recursive ? SearchOption.AllDirectories : SearchOption.TopDirectoryOnly))
+          ParseFile(fileName);
+      }
+      else
+      {
+        // TODO: Wrong source path.
       }
     }
